# Use codegen to speed up getstate and restore