Instead of doing inspection on the members with every call to getstate/restorestate this generates an "inlined" function specific to the class.  WIP as this still needs more testing and implemented for the SQLModel.

This leads to about a 2x improvement for getstate and and about a 20% overall gain in pydantics benchmark.

Benchmarks from pydantic's test case (with atomdb added here https://github.com/frmdstryr/pydantic/tree/add-atom-db-benchmark)
```
testing pydantic, atom, attrs + cattrs, valideer, 5 times each
          pydantic (1/5) time=0.369s, success=47.95%
          pydantic (2/5) time=0.367s, success=47.95%
          pydantic (3/5) time=0.370s, success=47.95%
          pydantic (4/5) time=0.388s, success=47.95%
          pydantic (5/5) time=0.395s, success=47.95%
          pydantic best=0.367s, avg=0.378s, stdev=0.013s

              atom (1/5) time=0.309s, success=47.95%
              atom (2/5) time=0.289s, success=47.95%
              atom (3/5) time=0.293s, success=47.95%
              atom (4/5) time=0.296s, success=47.95%
              atom (5/5) time=0.306s, success=47.95%
              atom best=0.289s, avg=0.299s, stdev=0.009s

    attrs + cattrs (1/5) time=0.297s, success=47.95%
    attrs + cattrs (2/5) time=0.298s, success=47.95%
    attrs + cattrs (3/5) time=0.298s, success=47.95%
    attrs + cattrs (4/5) time=0.300s, success=47.95%
    attrs + cattrs (5/5) time=0.300s, success=47.95%
    attrs + cattrs best=0.297s, avg=0.299s, stdev=0.001s

          valideer (1/5) time=0.397s, success=47.95%
          valideer (2/5) time=0.417s, success=47.95%
          valideer (3/5) time=0.408s, success=47.95%
          valideer (4/5) time=0.828s, success=47.95%
          valideer (5/5) time=0.484s, success=47.95%
          valideer best=0.397s, avg=0.507s, stdev=0.183s

          pydantic best=61.173μs/iter avg=63.000μs/iter stdev=2.125μs/iter version=1.9.0
              atom best=48.121μs/iter avg=49.765μs/iter stdev=1.433μs/iter version=0.7.0
    attrs + cattrs best=49.576μs/iter avg=49.795μs/iter stdev=0.195μs/iter version=21.4.0
          valideer best=66.176μs/iter avg=84.462μs/iter stdev=30.477μs/iter version=0.4.2


```